+++ CHANGELOG.md
@@ -6,12 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.10.1] 2022-04-13
+
 ### Changed
 - Set log level to DEBUG in dev (#952)
 - Remove Orchestrator{Aggregate, CompositeTrain, Test, Train}TaskSerializer and use orchestrator client directly (#920)
+- Handle error values in mapping functions (#945)
 
 ### Fixed
-- Handle error values in mapping functions (#945)
 - Handle disabled model in resync by making address not mandatory (#948)
 
 ## [0.10.0] 2022-04-11